+++ src/components/tmNotifications/index.js
@@ -1,0 +1,5 @@
+import tmNotifications from "./tmNotifications.vue"
+
+export default function install(Vue) {
+  Vue.component("tm-notifications", tmNotifications)
+}