+++ src/index.js
@@ -1,11 +1,13 @@
 import TmHelloWorld from "./components/tmHelloWorld"
 import TmField from "./components/tmField"
 import TmPageHeader from "./components/tmPageHeader"
+import TmNotifications from "./components/tmNotifications"
 
 const TendermintComponents = {
   TmHelloWorld,
   TmField,
-  TmPageHeader
+  TmPageHeader,
+  TmNotifications
 }
 
 TendermintComponents.install = Vue =>
@@ -23,3 +25,5 @@ export { default as TmField } from "./components/tmField/tmField.vue"
 export {
   default as TmPageHeader
 } from "./components/tmPageHeader/tmPageHeader.vue"
+  default as TmNotifications
+} from "./components/tmNotifications/tmNotifications.vue"